+++ test/spec/validator/phone.js
@@ -6,6 +6,7 @@ describe('phone', function() {
                     '<select class="form-control" name="country">',
                         '<option value="BR">Brazil</option>',
                         '<option value="CN">China</option>',
+                        '<option value="CZ">Czech Republic</option>',
                         '<option value="DK">Denmark</option>',
                         '<option value="ES">Spain</option>',
                         '<option value="FR">France</option>',
@@ -14,6 +15,7 @@ describe('phone', function() {
                         '<option value="PK">Pakistan</option>',
                         '<option value="RO">Romania</option>',
                         '<option value="RU">Russia</option>',
+                        '<option value="SK">Slovakia</option>',
                         '<option value="TH">Thailand</option>',
                         '<option value="US">USA</option>',
                         '<option value="VE">Venezuela</option>',
@@ -97,6 +99,35 @@ describe('phone', function() {
             this.$phone.val(validSamples[i]);
             this.bv.validate();
             expect(this.bv.isValid()).toBeTruthy();
+        }
+    });
+
+    it('Czech Republic phone number', function() {
+        this.bv.updateOption('phone', 'phone', 'country', 'CZ');
+
+        // Valid samples
+        var validSamples = [
+            '00420123456789', '00420 123456789', '00420 123 456 789', '00 420 123 456 789',
+            '+420123456789', '+420 123456789', '+420 123 456 789', '123456789', '123 456 789'
+        ];
+        for (var i in validSamples) {
+            this.bv.resetForm();
+            this.$phone.val(validSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toBeTruthy();
+        }
+
+        // Invalid samples
+        var invalidSamples = [
+            '420123456789', '420 123456789', '420 123 456 789', '00421123456789', '00421 123456789',
+            '00421 123 456 789', '00 421 123 456 789', '+421123456789', '+421 123456789',
+            '+421 123 456 789'
+        ];
+        for (i in invalidSamples) {
+            this.bv.resetForm();
+            this.$phone.val(invalidSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toEqual(false);
         }
     });
 
@@ -270,6 +301,35 @@ describe('phone', function() {
             this.$phone.val(validSamples[i]);
             this.bv.validate();
             expect(this.bv.isValid()).toBeTruthy();
+        }
+    });
+
+    it('Slovakia phone number', function() {
+        this.bv.updateOption('phone', 'phone', 'country', 'SK');
+
+        // Valid samples
+        var validSamples = [
+            '00420123456789', '00420 123456789', '00420 123 456 789', '00 420 123 456 789',
+            '+420123456789', '+420 123456789', '+420 123 456 789', '123456789', '123 456 789'
+        ];
+        for (var i in validSamples) {
+            this.bv.resetForm();
+            this.$phone.val(validSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toBeTruthy();
+        }
+
+        // Invalid samples
+        var invalidSamples = [
+            '420123456789', '420 123456789', '420 123 456 789', '00421123456789', '00421 123456789',
+            '00421 123 456 789', '00 421 123 456 789', '+421123456789', '+421 123456789',
+            '+421 123 456 789'
+        ];
+        for (i in invalidSamples) {
+            this.bv.resetForm();
+            this.$phone.val(invalidSamples[i]);
+            this.bv.validate();
+            expect(this.bv.isValid()).toEqual(false);
         }
     });
 });